buildsys: remove useless crap
The buildsystem was insanely complex for what is essentially a trivial
build. The cruft that deals with WIN32 has been removed because it just
adds bloat to serve a ridiculous use case. That allows us to use
GNUInstallDirs to sanitize the installation process. The weird and
useless ALVAR_PACKAGE logic has been removed and replaced with a sane
default installation that actually installs things where they belong.

The whole set of changes should simplify things significantly for
distribution package maintainers since we are now a lot closer to a
plain and simple cmake build.